+++ src/graphql/fetchConsentManagerId.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { GraphQLClient } from 'graphql-request';
 import {
   ConsentPrecedenceOption,
@@ -21,12 +20,10 @@ import {
   FETCH_CONSENT_MANAGER,
   EXPERIENCES,
   PURPOSES,
-  CONSENT_PARTITIONS,
   CONSENT_MANAGER_ANALYTICS_DATA,
   FETCH_CONSENT_MANAGER_THEME,
 } from './gqls';
 import { makeGraphQLRequest } from './makeGraphQLRequest';
-import { ConsentPartition } from '../codecs';
 
 export interface ConsentManager {
   /** ID of consent manager */
@@ -117,6 +114,8 @@ export interface ConsentPurpose {
   id: string;
   /** Name of purpose */
   name: string;
+  /** Tracking type of purpose */
+  trackingType: string;
 }
 
 /**
@@ -170,11 +169,15 @@ export interface ConsentExperience {
   purposes: {
     /** Name of purpose */
     name: string;
+    /** Purpose slug */
+    trackingType: string;
   }[];
   /** Purposes that are opted out by default in a particular experience */
   optedOutPurposes: {
     /** Name of purpose */
     name: string;
+    /** Purpose slug */
+    trackingType: string;
   }[];
   /**
    * Browser languages that define this regional experience
@@ -318,40 +321,3 @@ export async function fetchConsentManagerTheme(
   });
   return theme;
 }
-
-/**
- * Fetch the list of consent manager partitions
- *
- * @param client - GraphQL client
- * @returns Consent manager ID in organization
- */
-export async function fetchConsentManagerPartitions(
-  client: GraphQLClient,
-): Promise<ConsentPartition[]> {
-  const partitions: ConsentPartition[] = [];
-  let offset = 0;
-
-  // Fetch all partitions
-  let shouldContinue = false;
-  do {
-    const {
-      consentPartitions: { nodes },
-      // eslint-disable-next-line no-await-in-loop
-    } = await makeGraphQLRequest<{
-      /** Consent experience */
-      consentPartitions: {
-        /** List */
-        nodes: ConsentPartition[];
-      };
-    }>(client, CONSENT_PARTITIONS, {
-      first: PAGE_SIZE,
-      offset,
-    });
-    partitions.push(...nodes);
-    offset += PAGE_SIZE;
-    shouldContinue = nodes.length === PAGE_SIZE;
-  } while (shouldContinue);
-
-  return partitions.sort((a, b) => a.name.localeCompare(b.name));
-}
-/* eslint-enable max-lines */